Rename 'datamodel' folder to 'model'
