+++ app/src/main/java/app/wefridge/wefridge/model/OwnerControllerInterface.kt
@@ -1,6 +1,5 @@
-package app.wefridge.wefridge.datamodel
+package app.wefridge.wefridge.model
 
-import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.DocumentReference
 
 interface OwnerControllerInterface {
